+++ src/parser/inline/math.test.ts
@@ -11,6 +11,8 @@ describe('Unit: parser/inline/math', () => {
       assert.deepStrictEqual(inspect(parser('$')), undefined);
       assert.deepStrictEqual(inspect(parser('$$')), undefined);
       assert.deepStrictEqual(inspect(parser('$ $')), undefined);
+      assert.deepStrictEqual(inspect(parser('$ a$')), undefined);
+      assert.deepStrictEqual(inspect(parser('$ a $')), undefined);
       assert.deepStrictEqual(inspect(parser('$\n$')), undefined);
       assert.deepStrictEqual(inspect(parser('$a\nb$')), undefined);
       assert.deepStrictEqual(inspect(parser('$a\\\nb$')), undefined);
@@ -26,11 +28,11 @@ describe('Unit: parser/inline/math', () => {
       assert.deepStrictEqual(inspect(parser('$a$')), [['<span class="math notranslate" data-src="$a$">$a$</span>'], '']);
       assert.deepStrictEqual(inspect(parser('$ab$')), [['<span class="math notranslate" data-src="$ab$">$ab$</span>'], '']);
       assert.deepStrictEqual(inspect(parser('$a$b')), [['<span class="math notranslate" data-src="$a$">$a$</span>'], 'b']);
+      assert.deepStrictEqual(inspect(parser('$a $')), [['<span class="math notranslate" data-src="$a $">$a $</span>'], '']);
       assert.deepStrictEqual(inspect(parser('$a b$')), [['<span class="math notranslate" data-src="$a b$">$a b$</span>'], '']);
       assert.deepStrictEqual(inspect(parser('$\\a$')), [['<span class="math notranslate" data-src="$\\a$">$\\a$</span>'], '']);
       assert.deepStrictEqual(inspect(parser('$\\$$')), [['<span class="math notranslate" data-src="$\\$$">$\\$$</span>'], '']);
       assert.deepStrictEqual(inspect(parser('$\\\\$')), [['<span class="math notranslate" data-src="$\\\\$">$\\\\$</span>'], '']);
-      assert.deepStrictEqual(inspect(parser('$ a $')), [['<span class="math notranslate" data-src="$ a $">$ a $</span>'], '']);
     });
 
     it('nest', () => {