+++ src/parser/inline/media.test.ts
@@ -29,6 +29,8 @@ describe('Unit: parser/inline/media', () => {
       assert.deepStrictEqual(inspect(parser('![](a\nb)')), undefined);
       assert.deepStrictEqual(inspect(parser('![](a\\\nb)')), undefined);
       assert.deepStrictEqual(inspect(parser('![]((\n))')), undefined);
+      assert.deepStrictEqual(inspect(parser('![ a]()')), undefined);
+      assert.deepStrictEqual(inspect(parser('![ a ]()')), undefined);
       assert.deepStrictEqual(inspect(parser('![a\nb](ab)')), undefined);
       assert.deepStrictEqual(inspect(parser('![a\\\nb](ab)')), undefined);
       assert.deepStrictEqual(inspect(parser('![*a\\](\nb*](ab)')), undefined);
@@ -47,11 +49,8 @@ describe('Unit: parser/inline/media', () => {
       assert.deepStrictEqual(inspect(parser('![](\\b)')), [['<img class="media" data-src="\\b" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('![](a b)')), [['<img class="media invalid" data-src="a" alt="">'], '']);
       assert.deepStrictEqual(inspect(parser('![a]()')), [['<img class="media" data-src="" alt="a">'], '']);
-      assert.deepStrictEqual(inspect(parser('![a](b)')), [['<img class="media" data-src="b" alt="a">'], '']);
-      assert.deepStrictEqual(inspect(parser('![a b](c)')), [['<img class="media" data-src="c" alt="a b">'], '']);
-      assert.deepStrictEqual(inspect(parser('![a ](b)')), [['<img class="media" data-src="b" alt="a">'], '']);
-      assert.deepStrictEqual(inspect(parser('![ a](b)')), [['<img class="media" data-src="b" alt="a">'], '']);
-      assert.deepStrictEqual(inspect(parser('![ a ](b)')), [['<img class="media" data-src="b" alt="a">'], '']);
+      assert.deepStrictEqual(inspect(parser('![a ]()')), [['<img class="media" data-src="" alt="a">'], '']);
+      assert.deepStrictEqual(inspect(parser('![a b]()')), [['<img class="media" data-src="" alt="a b">'], '']);
     });
 
     it('nest', () => {