+++ src/parser/inline/strong.test.ts
@@ -13,6 +13,8 @@ describe('Unit: parser/inline/strong', () => {
       assert.deepStrictEqual(inspect(parser('***')), undefined);
       assert.deepStrictEqual(inspect(parser('****')), undefined);
       assert.deepStrictEqual(inspect(parser('** **')), undefined);
+      assert.deepStrictEqual(inspect(parser('** a**')), undefined);
+      assert.deepStrictEqual(inspect(parser('** a **')), undefined);
       assert.deepStrictEqual(inspect(parser('**\n**')), undefined);
       assert.deepStrictEqual(inspect(parser('**<wbr>**')), undefined);
       assert.deepStrictEqual(inspect(parser('****a****')), undefined);
@@ -21,6 +23,7 @@ describe('Unit: parser/inline/strong', () => {
 
     it('basic', () => {
       assert.deepStrictEqual(inspect(parser('**a**')), [['<strong>a</strong>'], '']);
+      assert.deepStrictEqual(inspect(parser('**a **')), [['<strong>a </strong>'], '']);
       assert.deepStrictEqual(inspect(parser('**ab**')), [['<strong>ab</strong>'], '']);
       assert.deepStrictEqual(inspect(parser('**a\nb**')), [['<strong>a<span class="linebreak"> <wbr></span>b</strong>'], '']);
       assert.deepStrictEqual(inspect(parser('**a\\\nb**')), [['<strong>a<br>b</strong>'], '']);